requirements.txt
matplotlib==3.10.3
numpy==2.3.1
pandas==2.0.3
PyYAML==6.0
PyYAML==6.0.2
seaborn==0.13.2
